% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulationRatePlot.R
\name{accumulationRatePlot}
\alias{accumulationRatePlot}
\title{Sediment Accumulation Rate Plot}
\usage{
accumulationRatePlot(
  basin,
  setting = c("valley", "interfluve"),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{setting}{a string, either "valley" or "interfluve".}

\item{xlim, ylim}{Optional setting to set the range of the x-axis and y-axis that is
displayed.}

\item{...}{Additional arguments to be passed to plot.}
}
\description{
Cross-section of sedimentary basin showing the relative rates of sediment accumulation in nonmarine portions of the basin.
}
\details{
Relative rates of sediment accumulation are shown in gray scale, from white (lowest rates of accumulation) to black (highest rates). Rates are not indicated for marine portions of the basin, which is shown in uniform gray.
}
\examples{
data(sedBasin)
accumulationRatePlot(sedBasin, setting="valley")

}
