% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHiatuses.R
\name{addHiatuses}
\alias{addHiatuses}
\title{Create a ternary plot}
\usage{
addHiatuses(
  basin,
  setting = c("valley", "interfluve"),
  col = "red",
  lwd = 2,
  ...
)
}
\arguments{
\item{\code{TRUE}}{}
}
\value{
\code{TRUE}
}
\description{
Creates a ternary plot from three variables that sum to 100\%, such as petrographic data in geology. Specifies the labels for the three apexes of the triangle, whether a grid should be shown, and the spacing of the grid lines
}
\examples{
ternaryPlot(myData, labels=("Q", "F", "L"))

}
