% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTimeLines.R
\name{addTimeLines}
\alias{addTimeLines}
\title{Time Lines}
\usage{
addTimeLines(basin, setting = c("valley", "interfluve"), timeLines, ...)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{setting}{a string, either "valley" or "interfluve".}

\item{timeLines}{a vector of times (in m.y.) for which time lines should be added.}

\item{...}{Optional arguments to style how the time lines are displayed.}
}
\description{
Adds time lines to a basin plot.
}
\details{
Time lines of any desired spacing are added to a basin plot.
}
\examples{
data(sedBasin)
plot(sedBasin)
addTimeLines(sedBasin, setting='valley', timeLines=seq(0.1, 2.9, 0.1))

}
