% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.R
\name{basin}
\alias{basin}
\alias{plot.basin}
\alias{print.basin}
\alias{summary.basin}
\title{subsidence: The subsidence function}
\usage{
basin(geometry, subsidence, eustasy, sediment, searchWidth = 10)

\method{plot}{basin}(
  x,
  setting = c("valley", "interfluve"),
  addLegend = TRUE,
  xlim = NULL,
  ylim = NULL,
  debugPlot = FALSE
)

\method{print}{basin}(x)

\method{summary}{basin}(x)
}
\arguments{
\item{geometry}{a geometry object}

\item{subsidence}{a subsidence object}

\item{eustasy}{a eustasy object}

\item{sediment}{a sediment object}

\item{searchWidth}{a distance (in km) for the distance over which an optimal shore is found. Generally should not be changed.}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Create a basin object
}
\details{
Creates a basin object, which records the history of sedimentation, elevation, and water depth across a sedimentary basin. Requires geometry, subsidence, eustasy, and sediment objects. Sedimentary basins can take several minutes to run.
}
\examples{
geom <- geometry(fallLineY=150, shoreX=200, deltaWidth=100, deltaToeY=-100, marginWidth=500, nonMarAlpha=0.5, marineAlpha=2.0, duration=3.0, timeStep=0.01)
subs <- subsidence(geometry=geom, startingLeft=0.0, startingRight=1.0)
eust <- eustasy(geometry=geom, netRise=30.0)	
sedi <- sediment(geometry=geom, startingVolume=120, netIncrease=120)
sedBasin <- basin(geometry=geom, subsidence=subs, eustasy=eust, sediment=sedi)

}
