% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bra_main_cities}
\alias{bra_main_cities}
\title{A tibble containing a list of the major Brazilian cities}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 133 rows and 5 columns.
}
\usage{
bra_main_cities
}
\description{
A tibble containing a list of the major Brazilian cities
}
\keyword{datasets}
