% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbAgeDepthPlot.R
\name{carbAgeDepthPlot}
\alias{carbAgeDepthPlot}
\title{Elevation Through Time for Carbonate Deposition}
\usage{
carbAgeDepthPlot(x, cex = 0.2)
}
\arguments{
\item{x}{an object of class 'column' generated by carbColumn().}

\item{cex}{specifies size of points.}
}
\value{
None.

geom <- geometry(fallLineY=150, shoreX=200, deltaWidth=100, deltaToeY=-100, marginWidth=500, resolution='medium', nonMarAlpha=0.5, marineAlpha=2.0, duration=12.0, timeStep=0.001)
subs <- subsidence(geometry=geom, startingLeft=0.0, startingRight=20.0, netChangeFactor=1.0)
eust <- eustasy(geometry=geom, period=3.0, amplitude=30, symmetry=0.5, phase='rising', shape=4, netRise=0.0)
depths <- c(0.0,  2.0,  5.0, 10.0, 40.0, 100.0)
rates  <- c(0.0, 40.0, 15.0, 10.0, 5.0, 2.0)
sedi <- carbSediment(depths, rates, lagTime=0.02, initialWaterDepth=2)
colu <- carbColumn(geometry=geom, subsidence=subs, eustasy=eust, carbSediment=sedi, locationKm=50.0)
carbAgeDepthPlot(colu)
}
\description{
Plot elevation of depositional surface through time for a stratigraphic column in shallow-marine carbonates.
}
\details{
Time points of carbonate deposition are shown in blue, and times of hiatus (when the carbonate platform is above sea level) are shown in red.
}
