% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbColumn.R
\name{carbColumn}
\alias{carbColumn}
\title{Stratigraphic Columns Simulating Shallow-marine Carbonate Deposition}
\usage{
carbColumn(geometry, subsidence, eustasy, carbSediment, locationKm = NULL)
}
\arguments{
\item{geometry}{an object of class \code{\link{geometry}}.}

\item{subsidence}{an object of class \code{\link{subsidence}}.}

\item{eustasy}{an object of class \code{\link{eustasy}}.}

\item{carbSediment}{an object of class \code{\link{carbSediment}}.}

\item{locationKm}{a number indicating where in the basin the column will be located,
specified as the distance (in km) from the left edge of the basin.}
}
\value{
returns an object of class \code{column}.

A column object consists of a data frame with five columns; each row corresponds to one sedimentary deposit. \code{modelTime} is the model time step in millions of years, \code{thickness} is the thickness of the deposit in meters, \code{stratPosition} is the height (in meters) within the stratigraphic column of the base of the deposit, \code{facies} is the facies of the deposit (carbonate, in this case), and \code{elevation} is the elevation (in meters above sea level) at which the deposit accumulated.
}
\description{
Create a stratigraphic column of shallow-marine carbonates..
}
\details{
Create a stratigraphic column of shallow-marine carbonates, based on a model similar to that of Read et al. (1986; Geology 14:107–110). Model includes tectonic subsidence, eustasy, and depth-dependent carbonate sedimentation rates, but not isostasy, compaction, or erosion.
}
\examples{
geom <- geometry(fallLineY=150, shoreX=200, deltaWidth=100, deltaToeY=-100,  
  marginWidth=500, resolution='medium', nonMarAlpha=0.5, marineAlpha=2.0,  
  duration=12.0, timeStep=0.001)
subs <- subsidence(geometry=geom, startingLeft=0.0, startingRight=20.0,  
  netChangeFactor=1.0)
eust <- eustasy(geometry=geom, period=3.0, amplitude=30, symmetry=0.5, phase='rising',  
  shape=4, netRise=0.0)
depths <- c(0.0,  2.0,  5.0, 10.0, 40.0, 100.0)
rates  <- c(0.0, 40.0, 15.0, 10.0, 5.0, 2.0)
sedi <- carbSediment(depths, rates, lagTime=0.02, initialWaterDepth=2)
colu <- carbColumn(geometry=geom, subsidence=subs, eustasy=eust, carbSediment=sedi,  
  locationKm=50.0)

}
