% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbSediment.R
\name{carbSediment}
\alias{carbSediment}
\title{Stratigraphic Columns}
\usage{
carbSediment(waterDepths, rates, lagTime = 0, initialWaterDepth = 0)
}
\arguments{
\item{waterDepths}{a vector containing the water depths (in m) at which sedimentation rate is specified. The first value must be 0.0, and subsequent values must increase monotonically.}

\item{rates}{a vector of the same length as waterDepths containing the corresponding sedimentation rates (in m / m.y.)}

\item{lagTime}{a numeric value giving the duration of time (in m.y.) that carbonate sedimentation will be paused following the flooding of a platform initially above sea level.}

\item{initialWaterDepth}{a numeric value giving the water depth at the beginning of the simulation of a stratigraphic column.}
}
\value{
returns a list with lagTime (a vector with a single value), initialWaterDepth (a vector with a single value), and productionCurve (a data frame with pairs of values of water depth and sedimentation rate).
}
\description{
Create Parameters for Carbonate Deposition.
}
\details{
Create the parameters that describe how carbonate sediments will be deposited. These include a series of pairs of water depths and sedimentation rates, lag time, and initial water depth. Carbonate sedimentation rates are given by the pairs, with sedimentation rates at intermediate water depths given by linear interpolation. Lag time describes the length of time that carbonate sedimentation will be paused following the flooding of a subaerially exposed platform. Lag time has been found to be necessary to create the ubiquitous flooding surfaces of shallow-marine carbonates. Initial water depth specifies the water depth at the beginning of a simulation, allowing a simulation to begin at any water depth.
}
\examples{
depths <- c(0.0,  2.0,  5.0, 10.0, 40.0, 100.0)
rates  <- c(0.0, 40.0, 15.0, 10.0, 5.0, 2.0)
sedi <- carbSediment(depths, rates, lagTime=0.02, initialWaterDepth=2)

}
