% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbThickness.R
\name{carbThickness}
\alias{carbThickness}
\title{Calculated Thickness of Carbonate Deposition}
\usage{
carbThickness(depth, timeStep, carbSed)
}
\arguments{
\item{depth}{a number indicating the current water depth (in m).}

\item{timeStep}{a number indicating the duration (in m.y.) of each simulated time step.}

\item{carbSed}{an object of class \code{carbSediment}.}
}
\value{
a numeric value of the thickness of carbonate sediment that was deposited.
}
\description{
Calculate the thickness of carbonate deposition during a single time step when creating a stratigraphic column.
}
\details{
Thickness of sediment is calculated.
}
\examples{
depths <- c(0.0,  2.0,  5.0, 10.0, 40.0, 100.0)
rates  <- c(0.0, 40.0, 15.0, 10.0, 5.0, 2.0)
carbSed <- carbSediment(depths, rates, lagTime=0.02, initialWaterDepth=2)
thickness <- carbThickness(depth=42, timeStep=0.01, carbSed=carbSed)

}
