% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_weights.R
\name{census_estimate_wgt}
\alias{census_estimate_wgt}
\title{Estimate aggregate tables with Census Microdata}
\usage{
census_estimate_wgt(dat, type = "apartment", operation = "rent")
}
\arguments{
\item{dat}{A `data.frame` or `data.table` imported with `census_import_census()`.}

\item{type}{One of `apartment` (default) or `house`.}

\item{operation}{One of `rent` (default) or `sale`.}
}
\value{
A named list
}
\description{
Estimates totals and averages by number of bedrooms and weighting areas using
Census household microdata. By default, only permanent private households with
non-zero income are included.
}
\details{
The default option includes only apartments (`v4002 == 13`) and rented units
(`v0201 == 3`). As houses we include both single detached houses and houses in
condominiums (`v4002 %in% c(11, 12)`). For sales we include all privately owned
houses, whether they have a mortgage or note (`v0201 %in% c(1, 2)`).

This function makes estimates faster by ignoring survey design and only providing
point estimates. That is, this function does not return estimates for standard
deviations. For standard deviations estimates consider using the `survey` package.
}
