% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_weights.R
\name{census_import_state}
\alias{census_import_state}
\title{Import household microdata from Census IBGE 2010}
\usage{
census_import_state(state, dir = NULL, label = FALSE)
}
\arguments{
\item{state}{Two letter abbreviation of the state. This option assumes the Census
microdata has been downloaded and unziped in the `dir` directory. Either a single
state or 'all'.}

\item{dir}{Path to the Census microdata directory. Defaults to `NULL` assuming
standard file names.}

\item{label}{Logical indicating if factor labels should be returned. Defaults
to `FALSE`.}
}
\value{
A `tibble`
}
\description{
Import household microdata from Census IBGE 2010
}
