% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnAgePlot.R
\name{columnAgePlot}
\alias{columnAgePlot}
\title{Column-age Plot}
\usage{
columnAgePlot(column, yAxisLabels = TRUE, ...)
}
\arguments{
\item{column}{an object of class \code{\link{column}}.}

\item{yAxisLabels}{a BOOL indicating whether labels should be included on the y-axis.}

\item{...}{Additional graphical parameters to be passed to the plot.}
}
\description{
A plot of model age through a stratigraphic column (a time-depth plot).
}
\details{
Plot shows the age of each horizon within a column.
}
\examples{
data(coluValley)
elevationHistoryPlot(coluValley)

}
