% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coluValley.R
\docType{data}
\name{coluValley}
\alias{coluValley}
\title{Example Stratigraphic Column through an Incised Valley}
\format{
An object of class \code{\link{column}}.
}
\usage{
data(coluValley)
}
\description{
Data generated by \code{column} in strataR: \code{column(sedBasin, locationKm=190.0,   setting='valley', pChannel=0.05, channelDepth=2.0)}
}
\examples{
data(coluValley)
plot(coluValley)
}
\keyword{datasets}
