% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{Completeness}
\usage{
completeness(occurrences, species)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}

\item{species}{an object of class \code{\link{species}}.}
}
\value{
A list of the total number species, the number of species that occurred at least once, the number species that occurred only once, and the number of species that occurred only twice.
}
\description{
Completeness of the fossil record.
}
\details{
\code{completeness} calculates the percentage of species that were preserved at least once in a column (completeness), the percentage of species that had only two occurrences in a column (two-timers), and the percentage of species that occurred only once (singletons).
}
\examples{
data(occu)
data(spec)
completeness(occurrences=occu, species=spec)

}
