% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevationBasinPlot.R
\name{elevationBasinPlot}
\alias{elevationBasinPlot}
\title{Elevation Basin Plot}
\usage{
elevationBasinPlot(
  basin,
  setting = c("valley", "interfluve"),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{setting}{a string, either "valley" or "interfluve".}

\item{xlim, ylim}{Optional setting to set the range of the x-axis and y-axis that is
displayed.}

\item{...}{Additional arguments to be passed to plot.}
}
\description{
Cross-section of sedimentary basin showing the elevations at the time of deposition in nonmarine portions of the basin.
}
\details{
Elevations are shown in gray scale, from white (sea level) to black (highest elevations). Elevations are not indicated for marine portions of the basin, which are shown in uniform gray.
}
\examples{
data(sedBasin)
elevationBasinPlot(sedBasin, setting="valley")

}
