% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevationDiversityPlot.R
\name{elevationDiversityPlot}
\alias{elevationDiversityPlot}
\title{Elevation-Diversity Plot}
\usage{
elevationDiversityPlot(
  species,
  minElevation = 0,
  maxElevation = 150,
  pCrit = 0.1
)
}
\arguments{
\item{species}{an object of class \code{\link{species}}.}

\item{minElevation, maxElevation}{a numeric value for the minimum and maximum elevation
to be considered, in meters above sea level.}

\item{pCrit}{a numeric value between 0 and 1 for the threshold probability of
collection for a species to be considered as occurring at a particular elevation.}
}
\description{
Plot of diversity vs. elevation for species.
}
\details{
Diagnostic plot to determine if diversity covaries with elevation; it shouldn't, but this might happen for stochastic reasons, especially if the number of species is small. Plot spans a specified range of elevations and records a species as present at a particular elevation if its probability of collection exceeds a minimum probability of collection. See \code{elevationDiversityResamplePlot()} for a version that adds confidence intervals.
}
\examples{
data(spec)
elevationDiversityPlot(species=spec, minElevation=20, maxElevation=200, pCrit=0.2)

}
