% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevationHistoryPlot.R
\name{elevationHistoryPlot}
\alias{elevationHistoryPlot}
\title{Elevation History Plot}
\usage{
elevationHistoryPlot(column, yAxisLabels = TRUE, ...)
}
\arguments{
\item{column}{an object of class \code{\link{column}}.}

\item{yAxisLabels}{a BOOL indicating whether labels should be included on the y-axis.}

\item{...}{Additional graphical parameters to be passed to the plot.}
}
\description{
A plot of the history of elevation in a stratigraphic column.
}
\details{
Plot shows how elevation above sea level (in m) preserved in a stratigraphic column changes within the column.
}
\examples{
data(coluValley)
elevationHistoryPlot(coluValley)

}
