% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevationTimePlot.R
\name{elevationTimePlot}
\alias{elevationTimePlot}
\title{Create a ternary plot}
\usage{
elevationTimePlot(basin, locationKm, setting = c("valley", "interfluve"), ...)
}
\arguments{
\item{\code{TRUE}}{}
}
\value{
\code{TRUE}
}
\description{
Creates a ternary plot from three variables that sum to 100\%, such as petrographic data in geology. Specifies the labels for the three apexes of the triangle, whether a grid should be shown, and the spacing of the grid lines
}
\examples{
ternaryPlot(myData, labels=("Q", "F", "L"))

}
