% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eust.R
\docType{data}
\name{eust}
\alias{eust}
\title{Example Eustatic Sea-Level History}
\format{
An object of class \code{\link{eustasy}}.
}
\usage{
data(eust)
}
\description{
Data generated by \code{eustasy} in strataR: \code{eustasy(geometry=geom, netRise=30.0)}
}
\examples{
data(eust)
plot(eust)
}
\keyword{datasets}
