% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLad.R
\name{fadLad}
\alias{fadLad}
\title{First and Last Occurrences (FADs and LADs)}
\usage{
fadLad(occurrences)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}
}
\value{
A data frame including the species ID of those that occur in a column, the stratigraphic position (in m) of their FAD and LAD, the model times (in m.y.) of the FAD and LAD, and the number of occurrences in the column.
}
\description{
\code{fadLad} returns the first and last occurrence of every species in a column.
}
\details{
Finds the first and last occurrence of every species in a stratigraphic column, reporting the stratigraphic position (in m) and the age of that horizon (in m.y.). Also reports the number of occurrences of each species.
}
\examples{
data(occu)
fadLad(occurrences=occu)

}
