% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLadElevationSedPlot.R
\name{fadLadElevationSedPlot}
\alias{fadLadElevationSedPlot}
\title{Plot FADs or LADs, Elevation, and Sedimentation Rate}
\usage{
fadLadElevationSedPlot(
  occurrences,
  column,
  type = c("fad", "lad"),
  sampleSpacing = 0.5,
  stratBin = 10,
  peakExtTimeMy = NA,
  extDurationMy = NA
)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}

\item{column}{an object of class \code{\link{column}}.}

\item{type}{a string of either "fad" or "lad" specifying which is to be plotted.}

\item{sampleSpacing}{a numeric value indicating the binning interval (in meters) for
tallying first or last occurrences in the stratigraphic column.}

\item{stratBin}{a numeric value indicating the stratigraphic interval (in meters) to be
binned for calculating sedimentation rates.}

\item{peakExtTimeMy}{a numeric value indicating when a mass extinction reached its peak
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}

\item{extDurationMy}{a numeric value indicating when the duration of a mass extinction
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}
}
\description{
Plot the number of FADs or LADs through a stratigraphic column, next to plots that show the change in elevation (meters above sea level) recorded in the column, and sedimentation rate.
}
\details{
Plot consists of three panels, combining \code{fadLadPlot}, \code{elevationHistoryPlot}, and \code{sedRateHistoryPlot}. The left panel shows the number of first and last occurrences through a stratigraphic column. If a mass extinction was simulated, its peak and its duration are plotted as well. The sampling interval for binning first and last occurrences in specified. The middle panel shows the changes in elevation preserved through the stratigraphic column, in meters above sea level. The right panel shows sedimentation rate (in m / m.y.) through the stratigraphic column. The interval (in meters) over which sedimentation rates are calculated is specified.
}
\examples{
data(occu)
data(coluValley)
fadLadElevationSedPlot(occurrences=occu, column=coluValley, type='lad', 
  sampleSpacing=0.5, stratBin=10)

}
