% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadLadPlot.R
\name{fadLadPlot}
\alias{fadLadPlot}
\title{Plot First and Last Occurrences (FADs and LADs)}
\usage{
fadLadPlot(
  occurrences,
  column,
  type = c("fad", "lad"),
  sampleSpacing = 0.5,
  xMax = NA,
  peakExtTimeMy = NA,
  extDurationMy = NA,
  extinctionColor = "red"
)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}

\item{column}{an object of class \code{\link{column}}.}

\item{type}{a string of either "fad" or "lad" specifying which is to be plotted.}

\item{sampleSpacing}{a numeric value indicating the binning interval (in meters) for
tallying first or last occurrences in the stratigraphic column.}

\item{xMax}{a numeric value for setting the largest number of first or last occurrences
to be shown. Use NA (the default) to not set a ceiling on this axis.}

\item{peakExtTimeMy}{a numeric value indicating when a mass extinction reached its peak
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}

\item{extDurationMy}{a numeric value indicating when the duration of a mass extinction
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}

\item{extinctionColor}{a color for displaying the window of extinction and the
stratigraphic position of peak extinction rates.}
}
\description{
Plot the number of first or last occurrences of specie through a stratigraphic column.
}
\details{
The number of first or last occurrences is shown, binned over a specified stratigraphic interval. If a mass extinction is simulated, the window of the mass extinction and the horizon that corresponds to peak extinction rates are also indicated.
}
\examples{
data(occu)
data(coluValley)
fadLadPlot(occurrences=occu, column=coluValley, type='fad', sampleSpacing=0.5)

}
