% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{geom}
\alias{geom}
\title{Example Geometry for Basin Simulation}
\format{
An object of class \code{\link{geometry}}.
}
\usage{
data(geom)
}
\description{
Data generated by \code{geometry} in strataR: \code{geometry(fallLineY=150, shoreX=200,   deltaWidth=100, deltaToeY=-100, marginWidth=600, nonMarAlpha=0.5, marineAlpha=2.0,   duration=3.0, timeStep=0.01)}
}
\examples{
data(geom)
summary(geom)
}
\keyword{datasets}
