% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_14_muni.R
\name{get_census_14_muni}
\alias{get_census_14_muni}
\title{Get Census Weights for the 14 cities}
\usage{
get_census_14_muni(geo = FALSE, dir = NULL)
}
\arguments{
\item{geo}{Logical indicating if Census weights should include a geometry column.
Defaults to `FALSE`}

\item{dir}{Directory path where the "14_municipios" data is stored}
}
\value{
Either a `tibble` (if `geo = FALSE`) or a spatial `sf` object
(if `geo = TRUE`).
}
\description{
IBGE changed the weighting areas for 14 cities. This function specifically
deals with this nuisance.
}
