% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wgt_area.R
\name{get_wgt_area}
\alias{get_wgt_area}
\title{Get the shape file of the Census microdata}
\usage{
get_wgt_area(city, dir = NULL)
}
\arguments{
\item{city}{7-digit IBGE code to filter the resulting shapefile. Defaults
to 'all'}
}
\value{
A `sf`
}
\description{
Get the shape file of the Census microdata
}
