% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexForPosition.R
\name{indexForPosition}
\alias{indexForPosition}
\title{Create a ternary plot}
\usage{
indexForPosition(positionKm, marginWidth, deltaX)
}
\arguments{
\item{\code{TRUE}}{}
}
\value{
\code{TRUE}
}
\description{
Creates a ternary plot from three variables that sum to 100\%, such as petrographic data in geology. Specifies the labels for the three apexes of the triangle, whether a grid should be shown, and the spacing of the grid lines
}
\examples{
ternaryPlot(myData, labels=("Q", "F", "L"))

}
