% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratedSedimentPlot.R
\name{integratedSedimentPlot}
\alias{integratedSedimentPlot}
\title{Integrated Sediment Plot}
\usage{
integratedSedimentPlot(basin)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}
}
\description{
Plot of the total sediment deposited through time.
}
\details{
Plot shows the actual volume of sediment deposited through time versus the targeted volume of sediment.
}
\examples{
data(sedBasin)
integratedSedimentPlot(sedBasin)

}
