% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcolu.R
\docType{data}
\name{marcolu}
\alias{marcolu}
\title{Example Stratigraphic Column that Includes Marine Strata}
\format{
An object of class \code{\link{column}}.
}
\usage{
data(marcolu)
}
\description{
Data generated by \code{geometry} in strataR: \code{column(sedBasin, locationKm=250.0,   setting='interfluve', pChannel=0.05, channelDepth=2.0)}
}
\examples{
data(marcolu)
plot(marcolu)
}
\keyword{datasets}
