% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineOccurrences.R
\name{marineOccurrences}
\alias{marineOccurrences}
\alias{plot.marineOccurrences}
\alias{summary.marineOccurrences}
\title{Marine Fossil Occurrences}
\usage{
marineOccurrences(column, marineSpecies, sampleSpacing = 0.5)

\method{plot}{marineOccurrences}(
  x,
  column,
  marineSpecies,
  occurrenceColor = "blue",
  rangeColor = "blue",
  extantColor = "gray",
  peakExtTimeMy = NA,
  extDurationMy = NA,
  extinctionColor = "red",
  orderedByLad = TRUE,
  yAxisLabels = TRUE,
  ...
)

\method{summary}{marineOccurrences}(object, ...)
}
\arguments{
\item{column}{an object of class \code{\link{column}}.}

\item{marineSpecies}{an object of class \code{\link{marineSpecies}}.}

\item{sampleSpacing}{a number indicating the stratigraphic interval (in m) at which the
column will be sampled for fossil species.}

\item{x, object}{an object of class \code{marineOccurrences}.}

\item{occurrenceColor, rangeColor, extantColor}{specify the color of the occurrence points, the preserved range of species, and the times in which the species was extant, including the time of origination and extinction.}

\item{peakExtTimeMy}{a numeric value (in m.y.) specifying the time of peak mass
extinction used; used in the occurrences plot.}

\item{extDurationMy}{a numeric value (in m.y.) specifying the duration of mass
extinction used; used in the occurrences plot.}

\item{extinctionColor}{a color specifying how a mass extinction interval should be
displayed on the occurrences plot.}

\item{orderedByLad}{a boolean specifying whether the occurrences should be ordered by
last occurrences (first occurrences if FALSE) on the occurrences plot.}

\item{yAxisLabels}{a boolean specifying whether y-axis labels should be displayed in
the occurrences plot.}

\item{...}{additional arguments to be passed.}
}
\value{
'marineOccurrences' returns a marineOccurrences object, which includes print, summary, and plot methods.

An occurrences object consists of a list of three items, each row corresponding to one occurrence of a fossil species. \code{modelTime} is the age (in m.y.) of the occurrence, \code{stratPosition} is the position in meters above the base of column in which the species occurred, and \code{speciesId} is the id number of the species that occurred. This species id can be cross-referenced with the species object to find the times of origination and extinction, as well as the ecological and preservational characteristics of that species.
}
\description{
Generate occurrences of marine fossils in a stratigraphic column.
}
\details{
Given a stratigraphic column and a set of species, \code{marineOccurrences} will sample that column for the occurrence of species at a specified sample spacing.
}
\examples{
data(marcolu)
data(marspec)
maroccu <- marineOccurrences(column=marcolu, marineSpecies=marspec, sampleSpacing=1.0)
summary(maroccu)
plot(maroccu, col=marcolu, marineSpecies=marspec)

}
