% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineSpecies.R
\name{marineSpecies}
\alias{marineSpecies}
\alias{plot.marineSpecies}
\alias{print.marineSpecies}
\alias{summary.marineSpecies}
\title{Marine Species Simulation}
\usage{
marineSpecies(
  timeStep,
  extRate = 0.25,
  startingSpecies = 200,
  durationMy = 10,
  peakExtFactor = 1,
  peakExtDurationMy = 1,
  peakExtTimeMy = 2,
  minPD,
  maxPD,
  meanDT,
  sdDT,
  meanLogPA,
  sdLogPA,
  maxPA
)

\method{plot}{marineSpecies}(x, ...)

\method{print}{marineSpecies}(x, ...)

\method{summary}{marineSpecies}(object, ...)
}
\arguments{
\item{timeStep}{the time step for simulating evolution, in m.y. Typically set to a
small value, such as 0.001 (1000 years).}

\item{extRate}{extinction rate, as a probability per m.y. Speciation rate is set to be
equal to the extinction rate, such that there is no forced change in diversity other
than what occurs through drift.}

\item{startingSpecies}{number of species at start of simulation.}

\item{durationMy}{duration of species simulation, in m.y.}

\item{peakExtFactor}{factor (multiplier) describing how much extinction rate will
increase at the peak of a mass extinction. Set to 1 to have no mass extinction.}

\item{peakExtDurationMy}{duration of mass extinction, in m.y.}

\item{peakExtTimeMy}{time at which mass extinction hits its peak rate, in m.y.
Extinction rate will climb linearly from background rates to peak rates through the
first half of the extinction window, then decrease linearly back to background rates
in the second half of the extinction window.}

\item{minPD}{minimum preferred water depth of species, in m below sea level.}

\item{maxPD}{maximum preferred water depth of species, in m below sea level.}

\item{meanDT}{mean water depth tolerance of species, in m.}

\item{sdDT}{standard deviation of water depth tolerance of species, in m.}

\item{meanLogPA}{mean base-10 logarithm of peak abundance of species.}

\item{sdLogPA}{standard deviation of the base-10 logarithm of peak abundance of
species.}

\item{maxPA}{maximum permitted peak abundance of species, as a percentage (e.g., set to
100 for the maximum peak abundance to be a 100\% probability of collection).}

\item{x, object}{an object of class \code{marineSpecies}.}

\item{...}{additional arguments to be passed to the plot.}
}
\value{
marineSpecies returns an object of class "marineSpecies", which as print, summary, and plot methods.

A marine species object consists of a list of 8 items, with each row corresponding to one species. \code{id} is a unique identifier corresponding to each species; this identifier starts at 1 and increases sequentially as each new species originates. \code{ancestor} corresponds to the id number of the species that gave rise to a species; its value is 0 for all of the seed species that are used to start the random-branching model. \code{origination} and \code{extinction} are the times of origination and extinction, in m.y. \code{PD} is the preferred depth (water depth, in m), that is, the water depth at which a species is most likely to occur. \code{DT} is the depth tolerance (in m) reflecting the standard deviation of the species response curve; larger values correspond to species found over a greater range of water depths. \code{PA} is the peak abundance, corresponding to the probability of occurrence (expressed as a percentage; i.e., 100 indicates a 1.0 probability) of a species at its preferred water depth.
}
\description{
Create a marine species object.
}
\details{
Creates a marine species object, which records the ancestor of a species, its times of origination and extinction, and its ecological characteristics: preferred water depth, water depth tolerance, and peak abundance. Evolution of species is treated as a random-branching process, which can be optionally be decreased during a mass extinction. A marine species object is designed for marine species; for nonmarine species, see species.
}
\examples{
spec <- marineSpecies(timeStep=0.001, extRate=0.25, startingSpecies=1000, 
  durationMy=5.0, minPD=0, maxPD=200, meanDT=10, sdDT=2, meanLogPA=log(25), 
  sdLogPA=log(5), maxPA=100)

}
