% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineWaterDepthDiversityPlot.R
\name{marineWaterDepthDiversityPlot}
\alias{marineWaterDepthDiversityPlot}
\title{Water Depth-Diversity Plot}
\usage{
marineWaterDepthDiversityPlot(
  marineSpecies,
  minWaterDepth = 0,
  maxWaterDepth = 150,
  pCrit = 0.1
)
}
\arguments{
\item{marineSpecies}{an object of class \code{\link{marineSpecies}}.}

\item{minWaterDepth, maxWaterDepth}{a numeric value for the minimum water depth to be
considered, in meters below sea level.}

\item{pCrit}{a numeric value between 0 and 1 for the threshold probability of
collection for a species to be considered as occurring at a particular water depth.}
}
\description{
Plot of diversity vs. water depth for marine species.
}
\details{
Diagnostic plot to determine if diversity covaries with water depth; it shouldn't, but this might happen for stochastic reasons, especially if the number of species is small. Plot spans a specified range of water depths and records a species as present at a particular water depth if its probability of collection exceeds a minimum probability of collection. See \code{marineWaterDepthDiversityResamplePlot()} for a version that adds confidence intervals.
}
\examples{
data(marspec)
marineWaterDepthDiversityPlot(marineSpecies=marspec, minWaterDepth=0, 
  maxWaterDepth=150, pCrit=0.1)

}
