% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineWaterDepthDiversityResamplePlot.R
\name{marineWaterDepthDiversityResamplePlot}
\alias{marineWaterDepthDiversityResamplePlot}
\title{Resampled Water Depth-Diversity Plot}
\usage{
marineWaterDepthDiversityResamplePlot(
  marineSpecies,
  minWaterDepth = 0,
  maxWaterDepth = 150,
  time = 0,
  numTrials = 100,
  conf = 0.5
)
}
\arguments{
\item{marineSpecies}{an object of class \code{\link{marineSpecies}}.}

\item{minWaterDepth, maxWaterDepth}{a numeric value for the minimum and maximum water
depth to be considered, in meters below sea level.}

\item{time}{a numeric value, in m.y., that limits analysis to species that originated
before a specified time and that went extinct after that same time. Generally left at
default to include species alive at beginning of simulation.}

\item{numTrials}{a numeric value for the number of resampling trials.}

\item{conf}{is a numeric value (0 to 1) indicating the confidence level used for
drawing the uncertainty envelope}
}
\description{
A plot of diversity vs. water depth based on resampling.
}
\details{
Diagnostic plot to determine if diversity covaries with water depth; it shouldn't, but this might happen for stochastic reasons, especially if the number of species is small. Plot spans a specified range of water depths and resamples to test if each species is present at a particular water depth. See \code{marineWaterDepthDiversityPlot()} for a faster version that lacks confidence intervals and is based on a fixed threshold probability of collection.
}
\examples{
data(marspec)
marineWaterDepthDiversityResamplePlot(marineSpecies=marspec, minWaterDepth=0, 
  maxWaterDepth=150, numTrials=100)

}
