% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maroccu.R
\docType{data}
\name{maroccu}
\alias{maroccu}
\title{Example Fossil Occurrences in Marine Strata}
\format{
An object of class \code{\link{marineOccurrences}}.
}
\usage{
data(maroccu)
}
\description{
Data generated by \code{marineOccurrences} in strataR: \code{marineOccurrences(marcolu, marspec,   sampleSpacing=0.5)}
}
\examples{
data(maroccu)
data(marcolu)
data(marspec)
plot(maroccu, column=marcolu, marineSpecies=marspec)
}
\keyword{datasets}
