% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marspec.R
\docType{data}
\name{marspec}
\alias{marspec}
\title{Example Marine Species}
\format{
An object of class \code{\link{marineSpecies}}.
}
\usage{
data(marspec)
}
\description{
Data generated by \code{marineSpecies} in strataR: \code{marineSpecies(timeStep=0.001,   extRate=0.25, startingSpecies=200, durationMy=5.0, minPD=0, maxPD=100, meanDT=10,   sdDT=2, meanLogPA=log(25), sdLogPA=log(5), maxPA=100)}
}
\examples{
data(marspec)
plot(marspec)
}
\keyword{datasets}
