% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu.R
\docType{data}
\name{occu}
\alias{occu}
\title{Example Fossil Occurrences in Nonmarine Strata}
\format{
An object of class \code{\link{occurrences}}.
}
\usage{
data(occu)
}
\description{
Data generated by \code{occurrences} in strataR: \code{occurrences(column=coluValley,   species=spec)}
}
\examples{
data(occu)
data(coluValley)
data(spec)
plot(occu, column=coluValley, species=spec)
}
\keyword{datasets}
