% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrenceColumnElevationPlot.R
\name{occurrenceColumnElevationPlot}
\alias{occurrenceColumnElevationPlot}
\title{Plot Fossil Occurrences and Elevation}
\usage{
occurrenceColumnElevationPlot(
  occurrences,
  column,
  species,
  occurrenceColor = "blue",
  rangeColor = "blue",
  extantColor = "gray",
  peakExtTimeMy = NA,
  extDurationMy = NA,
  orderedByLad = TRUE
)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}

\item{column}{an object of class \code{\link{column}}.}

\item{species}{an object of class \code{\link{species}}.}

\item{occurrenceColor, rangeColor, extantColor}{specify the color of the occurrence points, the preserved range of species, and the times in which the species was extant, including the time of origination and extinction.}

\item{peakExtTimeMy}{a numeric value indicating when a mass extinction reached its peak
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}

\item{extDurationMy}{a numeric value indicating when the duration of a mass extinction
(in m.y.), if one was simulated. Use NA (the default) if no extinction was simulated.}

\item{orderedByLad}{a boolean indicating how species will be sorted on the occurrence
chart. If TRUE, they are ordered by their last occurrences; if FALSE, they are
ordered by their first occurrences.}
}
\description{
Plot a stratigraphic column, a chart showing fossil occurrences and ranges, and elevation through the column.
}
\details{
Produces a three panel plot. The left plot is the stratigraphic column, produced with \code{plot.column}. Yellow indicates channel deposits, green indicates floodplain deposits, and tan corresponds to marine deposits. Hiatuses are shown with a red line.
The middle plot shows fossil occurrences and ranges, produced with \code{plot.occurrences}. Occurrences are shown with solid circles, except for singletons (species occurring only once), which are shown with open circles. Crosses indicate horizons that correspond to the times of origination and extinction. Blue lines indicate the preserved range of a species (i.e., connecting the first and last occurrence), and gray lines indicate the interval of which a species was extant (i.e., between the horizons corresponding to the times of origination and extinction. If a mass extinction was simulated, the stratigraphic interval corresponding to the time of extinction and its peak will be indicated on the stratigraphic column.
The right plot shows the changes in elevation (meters above sea level) preserved in the column. It is generated with \code{elevationHistoryPlot}.
}
\examples{
data(occu)
data(coluValley)
data(spec)
occurrenceColumnElevationPlot(occurrences=occu, column=coluValley, species=spec)

}
