% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{occurrences}
\alias{occurrences}
\alias{plot.occurrences}
\alias{print.occurrences}
\alias{summary.occurrences}
\title{Fossil Occurrences}
\usage{
occurrences(column, species, sampleSpacing = 0.5)

\method{plot}{occurrences}(
  x,
  column,
  species,
  peakExtTimeMy = NA,
  extDurationMy = NA,
  extinctionColor = "red",
  orderedByLad = TRUE,
  yAxisLabels = TRUE,
  ...
)

\method{print}{occurrences}(x, ...)

\method{summary}{occurrences}(object, ...)
}
\arguments{
\item{column}{an object of class \code{\link{column}}.}

\item{species}{an object of class \code{\link{species}}.}

\item{sampleSpacing}{a number indicating the stratigraphic interval (in m) at which the
column will be sampled for fossil species.}

\item{x, object}{an object of class \code{occurrences}.}

\item{peakExtTimeMy}{a numeric value (in m.y.) specifying the time of peak mass
extinction used; used in the occurrences plot.}

\item{extDurationMy}{a numeric value (in m.y.) specifying the duration of mass
extinction used; used in the occurrences plot.}

\item{extinctionColor}{a color specifying how a mass extinction interval should be
displayed on the occurrences plot.}

\item{orderedByLad}{a boolean specifying whether the occurrences should be ordered by
last occurrences (first occurrences if FALSE) on the occurrences plot.}

\item{yAxisLabels}{a boolean specifying whether y-axis labels should be displayed in
the occurrences plot.}

\item{...}{additional arguments to be passed.}
}
\value{
'occurrences' returns an occurrences object, which includes print, summary, and plot methods.

An occurrences object consists of a list of three items, each row corresponding to one occurrence of a fossil species. \code{modelTime} is the age (in m.y.) of the occurrence, \code{stratPosition} is the position in meters above the base of column in which the species occurred, and \code{speciesId} is the id number of the species that occurred. This species id can be cross-referenced with the species object to find the times of origination and extinction, as well as the ecological and preservational characteristics of that species.
}
\description{
Generate occurrences of non-marine fossils in a stratigraphic column.
}
\details{
Given a stratigraphic column and a set of species, \code{occurrences} will sample that column for the occurrence of species at a specified sample spacing.
}
\examples{
data(coluValley)
data(spec)
occurrences(column=coluValley, species=spec, sampleSpacing=1.0)

}
