% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeOffset.R
\name{rangeOffset}
\alias{rangeOffset}
\title{Range Offset}
\usage{
rangeOffset(occurrences, species)
}
\arguments{
\item{occurrences}{an object of class \code{\link{occurrences}}.}

\item{species}{an object of class \code{\link{species}}.}
}
\value{
a data frame listing the species that occurred in a stratigraphic column, the stratigraphic positions (in meters) of their first and last occurrences, and the range offsets (in m.y.) of their first and last occurrences.
}
\description{
Calculates range offset, the difference (in m.y.) between the time of origination and the age of the first occurrence, or between the time of extinction and the age of the last occurrence.
}
\details{
The first and last occurrence of a fossil species seldom corresponds to the times of origination and extinction. The mismatch is called range offset.
}
\examples{
data(occu)
data(spec)
rangeOffset(occurrences=occu, species=spec)

}
