% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedBasin.R
\docType{data}
\name{sedBasin}
\alias{sedBasin}
\title{Example Sedimentary Basin}
\format{
An object of class \code{\link{basin}}.
}
\usage{
data(sedBasin)
}
\description{
Data generated by \code{basin} in strataR: \code{basin(geometry=geom, subsidence=subs,   eustasy=eust, sediment=sedi)}
}
\examples{
data(sedBasin)
plot(sedBasin)
}
\keyword{datasets}
