% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedi.R
\docType{data}
\name{sedi}
\alias{sedi}
\title{Example Sediment Flux to a Basin}
\format{
An object of class \code{\link{sediment}}.
}
\usage{
data(sedi)
}
\description{
Data generated by \code{sediment} in strataR: \code{sediment(geometry=geom, startingVolume=60)}
}
\examples{
data(sedi)
plot(sedi)
}
\keyword{datasets}
