% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedRateHistoryPlot.R
\name{sedRateHistoryPlot}
\alias{sedRateHistoryPlot}
\title{Sedimentation Rate History Plot}
\usage{
sedRateHistoryPlot(column, stratBin = 10, yAxisLabels = TRUE, ...)
}
\arguments{
\item{column}{an object of class \code{\link{column}}.}

\item{stratBin}{a number (in m) indicating the stratigraphic thickness over which
average sedimentation rate should be calculated.}

\item{yAxisLabels}{a BOOL indicating whether labels should be included on the y-axis.}

\item{...}{Additional graphical parameters to be passed to the plot.}
}
\description{
Plot the history of sedimentation rate in a stratigraphic column.
}
\details{
Plots the sedimentation rates within a stratigraphic column, with sedimentation rates calculated over intervals of a given thickness.
}
\examples{
data(coluValley)
sedRateHistoryPlot(column=coluValley, stratBin=5)

}
