% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreAccommodationPlot.R
\name{shoreAccommodationPlot}
\alias{shoreAccommodationPlot}
\title{Shoreline Accommodation Plot}
\usage{
shoreAccommodationPlot(basin)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}
}
\description{
Plot of the rate of accommodation at the shoreline through time.
}
\details{
Shows the rate of accommodation (subsidence plus sea-level rise) at the shoreline through time. This will vary as rates of accommodation and eustasy vary through time, as well as with lateral shifts in the shoreline.
}
\examples{
data(sedBasin)
shoreAccommodationPlot(sedBasin)

}
