% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\docType{data}
\name{spec}
\alias{spec}
\title{Example Nonmarine Species}
\format{
An object of class \code{\link{species}}.
}
\usage{
data(spec)
}
\description{
Data generated by \code{species} in strataR: \code{species(timeStep=0.001, extRate=0.25,   startingSpecies=1000, durationMy=5.0, minPE=0, maxPE=200, meanET=10, sdET=2,   meanLogPA=log(25), sdLogPA=log(5), maxPA=100)}
}
\examples{
data(spec)
plot(spec)
}
\keyword{datasets}
