% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataR-package.R
\docType{package}
\name{strataR-package}
\alias{strataR}
\alias{strataR-package}
\title{strataR: Simulate Sedimentary Basins and Their Fossil Record}
\description{
strataR simulates the sedimentary fill of basins, based on inputs of tectonic subsidence, eustasy, sedimentation, and basin geometry. Stratigraphic columns can be simulated from anywhere in the basin, along with their fossil record.
}
\author{
\strong{Maintainer}: Steven Holland \email{stratum@uga.edu} (\href{https://orcid.org/0000-0002-4527-9370}{ORCID})

}
\keyword{internal}
