% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataR.R
\docType{package}
\name{strataR}
\alias{strataR}
\title{strataR A package for simulating sedimentary basins and their fossil records}
\description{
The package provides functions for creating sedimentary basins, stratigraphic columns from those basins, nonmarine and marine species that lived in those basins, and occurrences of fossils in stratigraphic columns. The package also provides functions for creating plots of all those objects.
}
