% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\docType{data}
\name{subs}
\alias{subs}
\title{Example Subsidence History}
\format{
An object of class \code{\link{subsidence}}.
}
\usage{
data(subs)
}
\description{
Data generated by \code{subsidence} in strataR: \code{subsidence(geometry=geom, startingLeft=0.0,   startingRight=1.0)}
}
\examples{
data(subs)
data(geom)
plot(subs, geometry=geom, type='lines')
}
\keyword{datasets}
