% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheelerPlot.R
\name{wheelerPlot}
\alias{wheelerPlot}
\title{Wheeler Plot}
\usage{
wheelerPlot(
  basin,
  setting = c("valley", "interfluve"),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{setting}{a string, either "valley" or "interfluve".}

\item{xlim, ylim}{Optional setting to set the range of the x-axis and y-axis that is
displayed.}

\item{...}{additional arguments to be passed.}
}
\description{
Plot of regions of sedimentation, non-deposition, and erosion throug time.
}
\details{
Plot is a chronostratigraphic diagram, also called a Wheeler plot, showing times and locations non-marine deposition (green), marine deposition (tan), subaerial erosion (the hiatus, black), and sediment starvation (white). Also shown are sediments that have been erosionally removed (the degradational vacuity, gray), as well as the location of the shore (blue). The combined hiatus and degradational vacuity represent the lacuna of an unconformity.
}
\examples{
data(sedBasin)
wheelerPlot(sedBasin, setting="valley")

}
