% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulationRatePlot.R
\name{accumulationRatePlot}
\alias{accumulationRatePlot}
\title{Create a ternary plot}
\usage{
accumulationRatePlot(
  basin,
  setting = c("valley", "interfluve"),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{\code{TRUE}}{}
}
\value{
\code{TRUE}
}
\description{
Creates a ternary plot from three variables that sum to 100\%, such as petrographic data in geology. Specifies the labels for the three apexes of the triangle, whether a grid should be shown, and the spacing of the grid lines
}
\examples{
ternaryPlot(myData, labels=("Q", "F", "L"))

}
