% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHiatuses.R
\name{addHiatuses}
\alias{addHiatuses}
\title{Hiatuses}
\usage{
addHiatuses(
  basin,
  setting = c("valley", "interfluve"),
  col = "red",
  lwd = 2,
  ...
)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{setting}{a string, either "valley" or "interfluve".}

\item{col}{the color of the hiatus line segments.}

\item{lwd}{the line weight of the hiatus line segments.}

\item{...}{Optional arguments to style how the hiatus line segments are displayed.}
}
\description{
Adds hiatuses to a basin plot.
}
\details{
Hiatuses are added to a basin plot. These may be subaerial unconformities
(typically within coastal plain deposits) or downlap surfaces (typically within
marine deposits).
}
\examples{
data(sedBasin)
plot(sedBasin)
addHiatuses(sedBasin, setting='valley')

}
