% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{column}
\alias{column}
\alias{plot.column}
\alias{print.column}
\alias{summary.column}
\title{Stratigraphic Columns}
\usage{
column(
  basin,
  locationKm,
  setting = c("valley", "interfluve"),
  pChannel,
  channelDepth
)

\method{plot}{column}(
  x,
  stratRange = c(floor(min(x$stratPosition)), ceiling(max(x$stratPosition))),
  axes = TRUE,
  ...
)

\method{print}{column}(x, ...)

\method{summary}{column}(object, ...)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{locationKm}{a number indicating where in the basin the column will be located,
specified as the distance (in km) from the left edge of the basin.}

\item{setting}{a string, either "valley" or "interfluve", specifying where the colulmn
will be located.}

\item{pChannel}{a number from 0 to 1 specifying the probability that a channel will be
present at any time step. Floodplain facies are deposited if a channel is not
present, so 1-pChannel can be considered the probability of floodplain facies.}

\item{channelDepth}{a number specifying the depth of a channel, in m.}

\item{x, object}{an object of class \code{column}.}

\item{stratRange}{a vector specifying (in m) the range of y-values on the column plot.}

\item{axes}{a boolean specifying whether to display the axes on the column plot.}

\item{...}{additional arguments to be passed to the column plot.}
}
\value{
\code{column} returns an object of class "column", which includes print and summary methods.

A column object consists of a list of five vectors, for the model time step (\code{modelTime}), the thickness of sediment accumulated, in meters (\code{thickness}), the stratigraphic position of each layer of sediment, in meters above the base of the column (\code{stratPosition}), the facies (floodplain or channel), and the elevation in meters above sea level at which deposition took place.
}
\description{
Create a stratigraphic column from a basin.
}
\details{
Create a stratigraphic column from anywhere in the basin by specifying the location of the column (the distance from the left edge of the basin) and by specifying whether the column is located along a valley or an interfluve. The probability of a channel occurring in any time step and the depth of the channel control control the relative proportions of chanel and floodplain facies.
}
\examples{
data(sedBasin)
column(sedBasin, locationKm=200, setting='valley', pChannel=0.1, channelDepth=2)

}
