% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{column}
\alias{column}
\alias{plot.column}
\alias{summary.column}
\title{Stratigraphic Columns}
\usage{
column(
  basin,
  locationKm,
  setting = c("valley", "interfluve"),
  pChannel,
  channelDepth
)

\method{plot}{column}(
  x,
  stratRange = c(floor(min(x$stratPosition)), ceiling(max(x$stratPosition))),
  axes = TRUE,
  ...
)

\method{summary}{column}(object, ...)
}
\arguments{
\item{basin}{an object of class \code{\link{basin}}.}

\item{locationKm}{a number indicating where in the basin the column will be located,
specified as the distance (in km) from the left edge of the basin.}

\item{setting}{a string, either "valley" or "interfluve", specifying where the column
will be located.}

\item{pChannel}{a number from 0 to 1 specifying the probability that a channel will be
present at any time step. Floodplain facies are deposited if a channel is not
present, so 1-pChannel can be considered the probability of floodplain facies.}

\item{channelDepth}{a number specifying the depth of a channel, in m.}

\item{x, object}{an object of class \code{column}.}

\item{stratRange}{a vector specifying (in m) the range of y-values on the column plot.}

\item{axes}{a boolean specifying whether to display the axes on the column plot.}

\item{...}{additional arguments to be passed to the column plot.}
}
\value{
\code{column} returns an object of class "column", which includes print and summary methods.

A column object consists of a data frame with five columns; each row corresponds to one sedimentary deposit. \code{modelTime} is the model time step in millions of years, \code{thickness} is the thickness of the deposit in meters, \code{stratPosition} is the height (in meters) within the stratigraphic column of the base of the deposit, \code{facies} is the facies of the deposit (floodplain, channel, or marine), and \code{elevation} is the elevation (in meters above sea level) at which the deposit accumulated.
}
\description{
Create a stratigraphic column from a basin.
}
\details{
Create a stratigraphic column from anywhere in the basin by specifying the location of the column (the distance from the left edge of the basin) and by specifying whether the column is located along a valley or an interfluve. The probability of a channel occurring in any time step and the depth of the channel control control the relative proportions of channel and floodplain facies.
}
\examples{
data(sedBasin)
colu <- column(sedBasin, locationKm=200, setting='valley', pChannel=0.1, channelDepth=2)
summary(colu)
plot(colu)

}
