% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geometry}
\alias{geometry}
\alias{plot.geometry}
\alias{print.geometry}
\alias{summary.geometry}
\title{geometry: The geometry function}
\usage{
geometry(
  fallLineY,
  shoreX,
  deltaWidth,
  deltaToeY,
  marginWidth,
  resolution = c("medium", "low", "high"),
  nonMarAlpha,
  marineAlpha,
  duration,
  timeStep
)

\method{plot}{geometry}(x)

\method{print}{geometry}(x)

\method{summary}{geometry}(x)
}
\arguments{
\item{fallLineY}{vertical elevation of the fall line, in meters. Corresponds to the edge of sedimentary basin, from which sediment is introduced.}

\item{shoreX}{distance of the initial shoreline from the fall line, in km.}

\item{deltaWidth}{width of the marine delta, in km.}

\item{deltaToeY}{initial water depth of the toe (base) of the marine delta, in m.}

\item{marginWidth}{width of the sedimentary basin, in km.}

\item{resolution}{a string (low, medium, or high) describing the spatial resolution of the model. In most cases, medium is sufficient.}

\item{nonMarAlpha}{a value from 0 to 1, describing the curvature of the coastal plain equilibrium profile.}

\item{duration}{duration of the basin simulation, in m.y.}

\item{timeStep}{duration of one time step in the basin simulation, in m.y.}

\item{marineAlphaa}{value from 0 to 1, describing the curvature of the marine equilibrium profile.}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Create a basin geometry object
}
\details{
Creates a basin geometry object, which is needed to create subsidence, eustasy, sedimentation, and basin objects. A geometry object defines the spatial and temporal characteristics of a basin simulation.
}
\examples{
geom <- geometry(fallLineY=150, shoreX=200, deltaWidth=100, deltaToeY=-100, marginWidth=500, nonMarAlpha=0.5, marineAlpha=2.0, duration=3.0, timeStep=0.01)

}
