% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment.R
\name{sediment}
\alias{sediment}
\alias{plot.sediment}
\alias{print.sediment}
\alias{summary.sediment}
\title{sediment: The sediment function}
\usage{
sediment(
  geometry,
  startingVolume = 60,
  netIncrease = 0,
  period = 1,
  amplitude = 0,
  symmetry = 0.5,
  phase = c("rising", "falling", "highPoint", "lowPoint"),
  shape = 0
)

\method{plot}{sediment}(x)

\method{print}{sediment}(x)

\method{summary}{sediment}(x)
}
\arguments{
\item{geometry}{a geometry object}

\item{startingVolume}{the starting sediment flux delivered at the left edge of the model, in m * km.}

\item{netIncrease}{the amount by which sediment influx increases beginning of the simulation to the end. Negative values indicate a decrease in sediment flux.}

\item{period}{the period of sediment flux cycles, in m.y.}

\item{amplitude}{the amplitude of sediment flux cycles, in m * km. The peak-to-peak sediment flux change is twice the amplitude.}

\item{symmetry}{a value ranging from 0 to 1 describing the symmetry of the sediment flux cycles. 0.5 is a classic cosine wave; smaller values have increasingly faster initial change in rates and slower subsequent changes in rates; larger values indicate the opposite.}

\item{phase}{a string with possible values of rising, falling, highPoint, and lowPoint describing the starting position of cyclic sediment flux.}

\item{shape}{a value of 0 or greater that describes the shape of the sediment flux cycle. 0 corresponds to a classic cosine curve, and larger values cause the curve to become increasingly squared-off.}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Create a sediment object
}
\details{
Creates a sediment object, which is needed to create a basin object. A sediment object describes how sediment flux to the basin changes through time
}
\examples{
geom <- geometry(fallLineY=150, shoreX=200, deltaWidth=100, deltaToeY=-100, marginWidth=500, nonMarAlpha=0.5, marineAlpha=2.0, duration=3.0, timeStep=0.01)
sedi <- sediment(geometry=geom, startingVolume=60)

}
