% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgt10}
\alias{wgt10}
\title{A shapefile of Census Weighting Areas for major cities in Brazil}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10993 rows and 18 columns.
}
\usage{
wgt10
}
\description{
A shapefile of Census Weighting Areas for major cities in Brazil
}
\keyword{datasets}
